broker/skills/templates/simpleSkill/requirements.txt
requests>=2.28.1
python-socketio[client]>=5.7.2
PyYAML>=6.0.1